feat: refactor BAHIS3 sync functions
